+++ routes/init.py
@@ -1,0 +1,7 @@
+from .user import router as user_router
+from .attraction import router as attraction_router
+from .mrt import router as mrt_router
+from .booking import router as booking_router
+from .order import router as order_router
+
+__all__ = ["user_router", "attraction_router", "mrt_router", "booking_router", "order_router"]
